requirements.txt
python-telegram-bot==21.5
pymongo==4.6.2